requirements.txt
Django==2.2.12
djangorestframework==3.11.0
django-allauth==0.40
django-rest_auth==0.9.5
django-dry-rest-permissions==1.1.0
django-simple-history==2.10.0
django-cors-headers==3.2.1
django-anymail==7.1.0
django-filter==2.2.0
coreapi==2.3.3
factory-boy==2.12.0
Pillow==7.1.1
djangorestframework-filters==0.11.1
python-decouple==3.3
django-storages==1.9.1
dj_database_url==0.5.0
zappa==0.51.0
psycopg2-binary==2.8.5
jsonfield==3.1.0
django-model-utils==4.0.0

# Documentation tools
mkdocs==1.1.2
mkdocs-material==5.2.3